TST: Adding tests for upload API and refactoring fixtures
It was non-trivial to add upload tests because of the database
setup required in the test environment as well. To avoid that, I
reworked the check if file exists to only require s3 and not a
connection to our database. In the process, I also had to rearrange
a lot of fixtures and environment variable use.