FIX: Add unique constraint to the status database (#308)
We were running into race conditions when adding multiple items
to the database. This will prevent that by raising an integrity
error when someone tries to commit a duplicate entry. It is a
partial index where we want to keep track of multiple FAILED entries,
but we only want one INPROGRESS or COMPLETED job to be in the table
at a time.