MNT: Refactor batch starter job kickoff (#309)
Reduce the nesting of loops and functions. Go through and try to
submit the job, if it doesn't succeed, continue on.

Rename the Status Tracking Table to Processing Job table.

MNT: Refactor batch job indexer updates

Add the unique table id to the job name for easier ability to
update the table in the subsequent updating lambdas.

Add the update logic directly into the updating lambda rather than a
separate file.