ENH: Add a DB synchronizer for our file database (#342)
This will keep s3 and our DB in sync with s3 being the source
of truth. Iterate through all items in s3 to get the keys, then
do a scan of the File Table to get all items in that table and
compare the two sets of items against one another. Then iterate
through adding / removing entries from the database table as necessary.